chore(#4860): namespace-cap relief + audit allowlist re-fingerprint
- tmux_watcher/two_message_panel.rs: inline #[cfg(test)] module moved
  verbatim to two_message_panel_tests.rs (#[path] sibling, precedent:
  turn_bridge/status_panel_tests.rs) — 780 -> 470 raw LoC, back under the
  tmux_watcher/** 700 cap.
- turn_bridge/status_panel.rs: completion-fallback helpers
  (persist_status_panel_completion_fallback_message_id + both fallback
  send transports) moved verbatim to status_panel/fallback.rs —
  740 -> 668 raw LoC, back under its 700 cap.
- audit_allowlist.toml: re-fingerprinted the two-message create send /
  footer-write keys (context shifted by the #4860 singleton sequence) and
  repointed the fallback send key to its new file (hash unchanged).
